Refactor `cmd` internal packages
