+++ internal/cmd/move/clients.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and IronCore contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package cmdutils
+package move
 
 import "sigs.k8s.io/controller-runtime/pkg/client"
